refactor: centralize hook execution logic in hookExecutor utility
Issue: The executeStatusHook method was defined in sessionManager.ts but conceptually belonged with other hook execution utilities.

Reason: Hook execution logic was scattered across multiple services, making it harder to maintain and test. All hook-related functionality should be centralized for better code organization.

Solution: Moved executeStatusHook from sessionManager.ts to hookExecutor.ts as a standalone function, following the same pattern as other hook utilities. This improves code modularity and makes hook execution logic easier to maintain and test.
